Clean up directory, Move csv to resources because of size
